Add files via upload
This is my first attempt at improving on the identical meaning checker from [https://huggingface.co/learn/nlp-course/chapter3/3?fw=pt]. While it is very good at spotting rewording e.g ('The French man bludgeoned a dinosaur with a rock'
vs 'A dinosaur was bludgeoned with a rock by a French man'), it does not spot synonyms e.g rock vs stone. Therefore, I made a dataset of synonyms with ChatGPT and also trained the AI on that. It did not work and lowered the score of identical prepositions that would previously get high scores.