Add files via upload
This my second attempt at the problem of identifying sentences which have identical meanings. My old approach of training the model on two separate datasets (a dataset of rephrased sentences and a dataset of synonyms) was inefficient as it led to the model unlearning some of what it had learned in the first dataset. This may be the reason that results varied depending on which dataset the model was trained from first. 

I therefore made a new dataset of sentences which have identical meanings by providing some examples of them to GPT-3 and then asking it to extend the list. While there were initially difficulties in using the dataset as for some reason, full stops in the dataset prevented the model from returning a loss, it does now work successfully. However, a much larger dataset is needed as the model is extremely confident that the sentences "the man was lying in the car" and "the man was reclining in the car" do not have the same meaning. 

I have also realised that I need to develop a standard evaluation procedure for my models, as I have just been thinking of sentences, throwing them in and then observing the results. Below are the results of doing that for the sym-checker-2 pipeline. While this kind of testing is useful for exploring which 'rules' the model is classifying by, it's not very good for comparing them. 

>>> test = pipeline("text-classification", "transformers-course/classifier1")
>>> test("this is a test" "this is not a test")
[{'label': 'LABEL_0', 'score': 0.9979117512702942}]
>>> test("the arabic man is in the car" "the arabic man is not in the automobile")
[{'label': 'LABEL_0', 'score': 0.995834469795227}]
>>> test("the french man bludgeoend a dinosaur with a rock" "a dinosaur was bludgeoned by the french man with a stone")
[{'label': 'LABEL_1', 'score': 0.7405192255973816}]
>>> test("the french man beat the dinosaur with a rock" "the dinosaur was beaten by a french man with a rock")
[{'label': 'LABEL_1', 'score': 0.9679585099220276}]
>>> test("the man was reclining in the car" "the man was sitting in the car")
[{'label': 'LABEL_0', 'score': 0.7766332030296326}]
>>> test("the man was reclining in the car" "the man was lying in the car")
[{'label': 'LABEL_0', 'score': 0.8808472752571106}]
>>> test("the man was happy in the car" "the man was sad in the car")
[{'label': 'LABEL_0', 'score': 0.9818387031555176}]
>>> test("the woman cleared dirt from the windscreen of the car" "the woman washed away dirt from the windscreen of the car")
[{'label': 'LABEL_1', 'score': 0.9903552532196045}]

Overall, it seems that the model has reasonable tolerance for spelling mistakes. It also seems to have recognised that the word 'not' inverts the meaning of a sentence as it provides high confidence and correct results when the word 'not' appears.