Add files via upload
I made the testing program and tried it out on the two models that I've made so far. 
Here are the results for the model trained on the GPT-3 dataset 
>TEST RUN 1
score
0.8571428571428571
certainty
0.9205681766782489
NEXT RUN BELOW
score
0.14285714285714285
certainty
0.8725155251366752

These mean that the model was 85% accurate in identifying sentences with the same meaning and was 92% sure. However, the model was 14% accurate at identifying sentences with opposite meanings and was 87% sure in these largely incorrect judgements. 

Here are the results from the first model which was trained on two datasets 
score
0.42857142857142855
certainty
0.8338272827012199
NEXT RUN BELOW
score
0.8571428571428571
certainty
0.8698342783110482

Overall, the first model performed significantly better at finding contradictory sentences. It returned about even results with high confidence for sentences with the same meaning. This overall suggests that more training is required. 

My next steps are to experiment with 
>refining datasets. It is possible that a larger or differently balanced dataset would return different results 
>training arguments. both of these models were trained using 'test-trainer' so there is possible scope for improvement here